pengulangan while fitur registrasi dan login
